+++ tests/commands/test_clusters.py
@@ -893,6 +893,90 @@ def test_cluster_expand_storage_fails(mock_request, capsys):
     assert "Some Error" in err_output
 
 
+times_change_product_operations_called = 0
+
+
+@pytest.mark.parametrize("status", ["SUCCEEDED", "FAILED", None])
+@mock.patch.object(Client, "request", return_value=({}, None))
+@mock.patch("time.sleep")
+def test_clusters_change_product(_mock_sleep, mock_request: mock.Mock, status):
+    product_name = "cr2"
+    cluster_id = gen_uuid()
+
+    def mock_call(*args, **kwargs):
+        if args[0] == RequestMethod.GET and "/operations/" in args[1]:
+            if status is None:
+                return None, None
+            global times_change_product_operations_called
+
+            if times_change_product_operations_called == 0:
+                times_change_product_operations_called += 1
+                return {"operations": [{"status": "SENT"}]}, None
+            return {"operations": [{"status": status}]}, None
+        return None, None
+
+    mock_request.side_effect = mock_call
+    call_command(
+        "croud",
+        "clusters",
+        "change-product",
+        "--cluster-id",
+        cluster_id,
+        "--product-name",
+        product_name,
+    )
+    print("requests ", mock_request.call_args_list)
+    assert_rest(
+        mock_request,
+        RequestMethod.PUT,
+        f"/api/v2/clusters/{cluster_id}/product/",
+        body={"product_name": product_name},
+        any_times=True,
+    )
+    assert_rest(
+        mock_request,
+        RequestMethod.GET,
+        f"/api/v2/clusters/{cluster_id}/operations/",
+        params={"type": "CHANGE_COMPUTE", "limit": 1},
+        any_times=True,
+    )
+    assert_rest(
+        mock_request,
+        RequestMethod.GET,
+        f"/api/v2/clusters/{cluster_id}/",
+        any_times=True,
+    )
+
+
+@mock.patch.object(Client, "request", return_value=(None, {}))
+def test_clusters_change_product_fails(mock_request, capsys):
+    product_name = "cr2"
+    cluster_id = gen_uuid()
+
+    def mock_call(*args, **kwargs):
+        return None, {"message": "Some Error"}
+
+    mock_request.side_effect = mock_call
+    call_command(
+        "croud",
+        "clusters",
+        "change-product",
+        "--cluster-id",
+        cluster_id,
+        "--product-name",
+        product_name,
+    )
+    assert_rest(
+        mock_request,
+        RequestMethod.PUT,
+        f"/api/v2/clusters/{cluster_id}/product/",
+        body={"product_name": product_name},
+    )
+
+    _, err_output = capsys.readouterr()
+    assert "Some Error" in err_output
+
+
 times_suspend_operations_called = 0