+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changes for croud
 Unreleased
 ==========
 
+- Added new subcommand ``clusters change-product`` that allows organizations and
+  project admins to change the cluster product.
+
 0.36.0 - 2022/07/27
 ===================